Created MovieSearchForm Componenet to put search form
